Major updates in mdp classe and added algorithm lauchers in mdp.algorithms package

+++ trunk/src/mdp/StateComparator.java
@@ -1,12 +1,13 @@
 package mdp;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Comparator;
 
 public class StateComparator implements Comparator{
 	   public int compare(Object o1, Object o2){
-		   Integer idState1= ((State) o1).getIdentifier();
-		   Integer idState2= ((State) o2).getIdentifier();
+		   BigInteger idState1= ((State) o1).getIdentifier();
+		   BigInteger idState2= ((State) o2).getIdentifier();
 		 /*  ArrayList<Boolean> l1= state1.getValues();
 		   ArrayList<Boolean> l2= state2.getValues();
 		   for (int i=0;i<l1.size();i++){
